+++ kwimage/__init__.py
@@ -12,7 +12,7 @@ from .im_alphablend import (ensure_alpha_channel, overlay_alpha_images,
 from .im_core import (atleast_3channels, ensure_float01, ensure_uint255,
                       make_channels_comparable, num_channels,)
 from .im_cv2 import (convert_colorspace, draw_boxes_on_image,
-                     draw_text_on_image, gaussian_patch, imscale,)
+                     draw_text_on_image, gaussian_patch, imscale, imresize,)
 from .im_demodata import (grab_test_image, grab_test_image_fpath,)
 from .im_io import (imread, imwrite,)
 from .im_runlen import (decode_run_length, encode_run_length, rle_translate,)
@@ -32,10 +32,11 @@ __all__ = ['Boxes', 'Coords', 'Detections', 'Heatmap', 'Mask', 'MaskList',
            'draw_text_on_image', 'encode_run_length', 'ensure_alpha_channel',
            'ensure_float01', 'ensure_uint255', 'gaussian_patch',
            'grab_test_image', 'grab_test_image_fpath', 'imread', 'imscale',
-           'imwrite', 'make_channels_comparable', 'non_max_supression',
-           'num_channels', 'overlay_alpha_images', 'overlay_alpha_layers',
-           'rle_translate', 'smooth_prob', 'stack_images', 'stack_images_grid',
-           'subpixel_accum', 'subpixel_align', 'subpixel_getvalue',
-           'subpixel_maximum', 'subpixel_minimum', 'subpixel_set',
-           'subpixel_setvalue', 'subpixel_slice', 'subpixel_translate',
-           'warp_points', 'warp_tensor']
+           'imresize', 'imwrite', 'make_channels_comparable',
+           'non_max_supression', 'num_channels', 'overlay_alpha_images',
+           'overlay_alpha_layers', 'rle_translate', 'smooth_prob',
+           'stack_images', 'stack_images_grid', 'subpixel_accum',
+           'subpixel_align', 'subpixel_getvalue', 'subpixel_maximum',
+           'subpixel_minimum', 'subpixel_set', 'subpixel_setvalue',
+           'subpixel_slice', 'subpixel_translate', 'warp_points',
+           'warp_tensor']